+++ v2/src/dar_backup/util.py
@@ -19,12 +19,12 @@ import sys
 import threading
 import traceback
 from datetime import datetime
+from dar_backup.config_settings import ConfigSettings
 
 from typing import NamedTuple, List
 
 logger=None
 secondary_logger=None   
-
 
 def setup_logging(log_file: str, command_output_log_file: str, log_level: str = "info", log_to_stdout: bool = False) -> logging.Logger:
     """
@@ -95,6 +95,47 @@ def get_logger(command_output_logger: bool = False) -> logging.Logger:
     global logger, secondary_logger
 
     return secondary_logger if command_output_logger else logger
+
+
+
+def requirements(type: str, config_setting: ConfigSettings):
+    """
+    Perform PREREQ or POSTREQ requirements.
+
+    Args:
+        type (str): The type of prereq (PREREQ, POSTREQ).
+        config_settings (ConfigSettings): An instance of the ConfigSettings class.
+
+    Raises:
+        RuntimeError: If a subprocess returns anything but zero.
+
+        subprocess.CalledProcessError: if CalledProcessError is raised in subprocess.run(), let it bobble up.
+    """
+    
+    if type is None or config_setting is None:
+        raise RuntimeError(f"requirements: 'type' or config_setting is None")
+
+    allowed_types = ['PREREQ', 'POSTREQ'] 
+    if type not in allowed_types:
+        raise RuntimeError(f"requirements: {type} not in: {allowed_types}")
+
+
+    logger.debug(f"Performing  {type}")
+    if type in config_setting.config:
+        for key in sorted(config_setting.config[type].keys()):
+            script = config_setting.config[type][key]
+            try:
+                result = subprocess.run(script, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True, shell=True, check=True)
+                logger.debug(f"{type} {key}: '{script}' run, return code: {result.returncode}")
+                logger.debug(f"{type} stdout:\n{result.stdout}")
+                if result.returncode != 0:
+                    logger.error(f"{type} stderr:\n{result.stderr}")
+                    raise RuntimeError(f"{type} {key}: '{script}' failed, return code: {result.returncode}")    
+            except subprocess.CalledProcessError as e:
+                logger.error(f"Error executing {key}: '{script}': {e}")
+                raise e
+
+
 
 
 class BackupError(Exception):